* Tools/ddbd/ddbd.m: Added -userMetadataForPath: that returns all the user-set metadata attributes (file annotations, etc.).
* Tools/ddbd/ddbd.m: 
* Tools/ddbd/MDModules/MDModuleAnnotations/MDModuleAnnotations.m: 
Changed "MDAnnotations" with "GSMDItemFinderComment" to be consistent with the other attribute names.

* Tools/ddbd/DDBPathsManager.m: 
In -setMetadata:ofType:forPath: post a distributed notification to permit to mdextractor to put the attribute also into the main db. 
Added -metadataForPath: to return all the user metadata associated with a path.

* GWMetadata/gmds/mdextractor/mdextractor.m:
* GWMetadata/gmds/mdextractor/updater.m:
Insert into the db also the user attributes returned by ddbd.


git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gworkspace/trunk@23697 72102866-910b-0410-8b05-ffd578937521