* FSNode/FSNBrowserCell.m: -setOpened: rewritten.
* GWMetadata/MDKit/SQLite.[hm]:
Added method to attach a database file.

* GWMetadata/gmds/dbschema.h:
Added sql statements to create the "updated" tables.
Added a "db_version" string.

* GWMetadata/gmds/mdextractor/updater.m:
When add or remove a path, inserts it into the "updated" or "removed" tables.
Sends notifications to let the live queries to update.

* GWMetadata/gmds/mdextractor/Extractors/AppExtractor/AppExtractor.m:
* GWMetadata/gmds/mdextractor/Extractors/JpegExtractor/JpegExtractor.m:
in -extractMetadataAtPath:withID:attributes: add a empty "words" NSSet to the dictionary passed to the mdextractor method. 

* GWMetadata/gmds/mdextractor/Extractors/HtmlExtractor/HtmlExtractor.m:
Don't return NO if the contents length is 0. 


git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gworkspace/trunk@24024 72102866-910b-0410-8b05-ffd578937521