* FSNode/FSNodeRepIcons.m: -resizedIcon:ofSize: Rewritten.
* GWorkspace/FileViewer/GWViewerSplit.m:
In -updateDiskSpaceInfo: don't call -drawWithFrame: if the frame is NSZeroRect.

* Tools/fswatcher/fswatcher-inotify.hm:
In -client:addWatcherForPath: check the value returned by inotify_add_watch() before using it in -initWithWatchedPath:watchDescriptor:fswatcher:.

* Tools/ddbd/ddbd.m:
-fileSystemDidChange: changed to be called directly by GWorkspace instead of using distributed notifications.

* GWorkspace/GWorkspace.m:
In -fileSystemDidChange: notify also ddbd. This to be sure that ddbd updates file annotations and other user-set metadata before mdextractor access it.



git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gworkspace/trunk@24488 72102866-910b-0410-8b05-ffd578937521